+++ lib/versao.ts
@@ -1,3 +1,3 @@
-export const VERSAO = '0.7.0';
+export const VERSAO = '0.8.0';
 
-export const VERSAO_NOME = 'HTTPS no domínio';
+export const VERSAO_NOME = 'WhatsApp recebendo';